requirements.txt
pip>=24.2

flake8>=7.3.0

openai>=1.0.0
chromadb >=1.0.20
python-dotenv>=1.0.0

unidecode>=1.4.0

fastapi
uvicorn

python-multipart>=0.0.20
sounddevice>=0.5.0
scipy>=1.16.1